backend/requirements.txt
appdirs==1.4.4
bcrypt==3.2.0
bidict==0.21.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudinary==1.24.0
cryptography==3.4.6
distlib==0.3.1
fastapi==0.60.1
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
flask-marshmallow==0.13.0
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
greenlet==1.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.7.1
oauthlib==3.0.1
pycparser==2.20
pydantic==1.6.1
PyJWT==1.4.2
PyMySQL==0.10.0
pyOpenSSL==19.0.0
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2021.1
queuelib==1.6.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.18
starlette==0.13.6
Twisted==21.2.0
twisted-iocpsupport==1.0.1
typing-extensions==3.10.0.0
urllib3==1.26.4
virtualenv==20.0.27
Werkzeug==1.0.1
requests-oauthlib==1.3.0
